edge-gateway/requirements.txt
# CAN Bus Interface
python-can==4.3.1
can-isotp==2.0.3

# MQTT and AWS IoT
paho-mqtt==1.6.1
awsiotsdk==1.20.0

# Protocol Decoders
udsoncan==1.21.1
python-obd==0.7.1

# Data Processing
pyyaml==6.0.1

# Database
sqlite3

# Utilities
python-dateutil==2.8.2
